+++ lib/ricardos_utils.dart
@@ -1,3 +1,4 @@
+/// A set of loosely related utilities and widgets.
 library ricardos_utils;
 
 export 'src/extensions.dart';