+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.4
+
+- Added `NumExtensions`:
+  - `toStringAsSigned`: Converts a number to a string with a sign
+  - `toFileSize`: Scales a number to a human-readable file size and returns a tuple with the scaled value and the unit
+  - `logBase`: Returns the logarithm of the number with a specified base
+  - `roundToFixed`: Rounds the number to a fixed number of decimal places
+
 ## 1.0.3
 
 - Fixed an issue with `Future<T>.thenBuild` for cases when `T` is nullable